requirements.txt
cdsapi==0.5.1
dask==2022.12.1
gdal==3.6.2
geopandas==0.12.2
h5netcdf==1.1.0
h5py==3.7.0
netcdf4==1.6.2
numpy==1.23.5
pandas==1.5.2
pyproj==3.4.1
pyyaml==6.0
rasterio==1.3.4
shapely==2.0.0
tqdm==4.64.1
xarray==2022.12.0
xclim==0.40.0
